+++ py/desitarget/gaiamatch.py
@@ -11,11 +11,19 @@ import os
 import sys
 import numpy as np
 import fitsio
+import requests
+import pickle
+from glob import glob
 from time import time
 import healpy as hp
+from os.path import basename
+from desitarget import io
 from desitarget.io import check_fitsio_version
+from desitarget.internal import sharedmem
+from desimodel.footprint import radec2pix
 from astropy.coordinates import SkyCoord
 from astropy import units as u
+from astropy.io import ascii
 
 # ADM set up the DESI default logger
 from desiutil.log import get_logger
@@ -26,27 +34,427 @@ start = time()
 
 # ADM the current data model for Gaia columns for READING from Gaia files
 ingaiadatamodel = np.array([], dtype=[
-            ('SOURCE_ID', '>i8'), ('RA', '>f8'), ('DEC', '>f8'),
+            ('SOURCE_ID', '>i8'), ('REF_CAT', 'S2'), ('RA', '>f8'), ('DEC', '>f8'),
             ('PHOT_G_MEAN_MAG', '>f4'), ('PHOT_G_MEAN_FLUX_OVER_ERROR', '>f4'),
             ('PHOT_BP_MEAN_MAG', '>f4'), ('PHOT_BP_MEAN_FLUX_OVER_ERROR', '>f4'),
             ('PHOT_RP_MEAN_MAG', '>f4'), ('PHOT_RP_MEAN_FLUX_OVER_ERROR', '>f4'),
+            ('PHOT_BP_RP_EXCESS_FACTOR', '>f4'),
             ('ASTROMETRIC_EXCESS_NOISE', '>f4'), ('DUPLICATED_SOURCE', '?'),
+            ('ASTROMETRIC_SIGMA5D_MAX', '>f4'), ('ASTROMETRIC_PARAMS_SOLVED', '>i1'),
             ('PARALLAX', '>f4'), ('PARALLAX_ERROR', '>f4'),
             ('PMRA', '>f4'), ('PMRA_ERROR', '>f4'),
-            ('PMDEC', '>f4'), ('PMDEC_ERROR', '>f4'),
+            ('PMDEC', '>f4'), ('PMDEC_ERROR', '>f4')
                                    ])
 
 # ADM the current data model for Gaia columns for WRITING to target files
 gaiadatamodel = np.array([], dtype=[
-            ('REF_ID', '>i8'), ('GAIA_RA', '>f8'), ('GAIA_DEC', '>f8'),
+            ('REF_ID', '>i8'), ('REF_CAT', 'S2'), ('GAIA_RA', '>f8'), ('GAIA_DEC', '>f8'),
             ('GAIA_PHOT_G_MEAN_MAG', '>f4'), ('GAIA_PHOT_G_MEAN_FLUX_OVER_ERROR', '>f4'),
             ('GAIA_PHOT_BP_MEAN_MAG', '>f4'), ('GAIA_PHOT_BP_MEAN_FLUX_OVER_ERROR', '>f4'),
             ('GAIA_PHOT_RP_MEAN_MAG', '>f4'), ('GAIA_PHOT_RP_MEAN_FLUX_OVER_ERROR', '>f4'),
+            ('GAIA_PHOT_BP_RP_EXCESS_FACTOR', '>f4'),
             ('GAIA_ASTROMETRIC_EXCESS_NOISE', '>f4'), ('GAIA_DUPLICATED_SOURCE', '?'),
+            ('GAIA_ASTROMETRIC_SIGMA5D_MAX', '>f4'), ('GAIA_ASTROMETRIC_PARAMS_SOLVED', '>i1'),
             ('PARALLAX', '>f4'), ('PARALLAX_IVAR', '>f4'),
             ('PMRA', '>f4'), ('PMRA_IVAR', '>f4'),
-            ('PMDEC', '>f4'), ('PMDEC_IVAR', '>f4'),
+            ('PMDEC', '>f4'), ('PMDEC_IVAR', '>f4')
                                    ])
+
+
+def _get_gaia_dir():
+    """Convenience function to grab the Gaia environment variable.
+
+    Returns
+    -------
+    :class:`str`
+        The directory stored in the $GAIA_DIR environment variable.
+    """
+    # ADM check that the $GAIA_DIR environment variable is set.
+    gaiadir = os.environ.get('GAIA_DIR')
+    if gaiadir is None:
+        msg = "Set $GAIA_DIR environment variable!"
+        log.critical(msg)
+        raise ValueError(msg)
+
+    return gaiadir
+
+
+def _get_gaia_nside():
+    """Grab the HEALPixel nside to be used throughout this module.
+
+    Returns
+    -------
+    :class:`int`
+        The HEALPixel nside number for Gaia file creation and retrieval.
+    """
+    nside = 32
+
+    return nside
+
+
+def scrape_gaia(url="http://cdn.gea.esac.esa.int/Gaia/gdr2/gaia_source/csv/", nfiletest=None):
+    """Retrieve the bulk CSV files released by the Gaia collaboration.
+
+    Parameters
+    ----------
+    url : :class:`str`
+        The web directory that hosts the archived Gaia CSV files.
+    nfiletest : :class:`int`, optional, defaults to ``None``
+        If an integer is sent, only retrieve this number of files, for testing.
+
+    Returns
+    -------
+    Nothing
+        But the archived Gaia CSV files are written to $GAIA_DIR/csv.
+
+    Notes
+    -----
+        - The environment variable $GAIA_DIR must be set.
+        - Runs in about 26 hours for 60,000 Gaia files.
+    """
+    # ADM check that the GAIA_DIR is set and retrieve it.
+    gaiadir = _get_gaia_dir()
+
+    # ADM construct the directory to which to write files.
+    csvdir = os.path.join(gaiadir, 'csv')
+    # ADM the directory better be empty for the wget!
+    if os.path.exists(csvdir):
+        if len(os.listdir(csvdir)) > 0:
+            msg = "{} should be empty to wget Gaia csv files!".format(csvdir)
+            log.critical(msg)
+            raise ValueError(msg)
+    # ADM make the directory, if needed.
+    else:
+        log.info('Making Gaia directory for storing CSV files')
+        os.makedirs(csvdir)
+
+    # ADM pull back the index.html from the url.
+    index = requests.get(url)
+
+    # ADM retrieve any file name that starts with GaiaSource.
+    # ADM the [1::2] pulls back just the odd lines from the split list.
+    filelist = index.text.split("GaiaSource")[1::2]
+
+    # ADM if nfiletest was passed, just work with that number of files.
+    test = nfiletest is not None
+    if test:
+        filelist = filelist[:nfiletest]
+    nfiles = len(filelist)
+
+    # ADM loop through the filelist.
+    t0 = time()
+    for nfile, fileinfo in enumerate(filelist):
+        # ADM make the wget command to retrieve the file and issue it.
+        cmd = 'wget -q {}/GaiaSource{} -P {}'.format(url, fileinfo[:-2], csvdir)
+        os.system(cmd)
+        if nfile % 100 == 0 or test:
+            elapsed = time() - t0
+            rate = nfile / elapsed
+            log.info(
+                '{}/{} files; {:.1f} files/sec; {:.1f} total mins elapsed'
+                .format(nfile+1, nfiles, rate, elapsed/60.)
+            )
+
+    log.info('Done...t={:.1f}s'.format(time()-t0))
+
+    return
+
+
+def gaia_csv_to_fits(numproc=4):
+    """Convert files in $GAIA_DIR/csv to files in $GAIA_DIR/fits.
+
+    Parameters
+    ----------
+    numproc : :class:`int`, optional, defaults to 4
+        The number of parallel processes to use.
+
+    Returns
+    -------
+    Nothing
+        But the archived Gaia CSV files in $GAIA_DIR/csv are converted
+        to FITS files in the directory $GAIA_DIR/fits. Also, a look-up
+        table is written to $GAIA_DIR/fits/hpx-to-files.pickle for which
+        each index is an nside=_get_gaia_nside(), nested scheme HEALPixel
+        and each entry is a list of the FITS files that touch that HEAPixel.
+
+    Notes
+    -----
+        - The environment variable $GAIA_DIR must be set.
+        - if numproc==1, use the serial code instead of the parallel code.
+        - Runs in 1-3 hours (depending on node) with numproc=32 for 60,000 files.
+    """
+    # ADM the resolution at which the Gaia HEALPix files should be stored.
+    nside = _get_gaia_nside()
+
+    # ADM check that the GAIA_DIR is set.
+    gaiadir = _get_gaia_dir()
+    log.info("running on {} processors".format(numproc))
+
+    # ADM construct the directories for reading/writing files.
+    csvdir = os.path.join(gaiadir, 'csv')
+    fitsdir = os.path.join(gaiadir, 'fits')
+
+    # ADM make sure the output directory is empty.
+    if os.path.exists(fitsdir):
+        if len(os.listdir(fitsdir)) > 0:
+            msg = "{} should be empty to make Gaia FITS files!".format(fitsdir)
+            log.critical(msg)
+            raise ValueError(msg)
+    # ADM make the output directory, if needed.
+    else:
+        log.info('Making Gaia directory for storing FITS files')
+        os.makedirs(fitsdir)
+
+    # ADM construct the list of input files.
+    infiles = glob("{}/*csv*".format(csvdir))
+    nfiles = len(infiles)
+
+    # ADM the critical function to run on every file.
+    def _write_gaia_fits(infile):
+        """read an input name for a csv file and write it to FITS"""
+        outbase = os.path.basename(infile)
+        outfilename = "{}.fits".format(outbase.split(".")[0])
+        outfile = os.path.join(fitsdir, outfilename)
+        fitstable = ascii.read(infile, format='csv')
+
+        # ADM need to convert 5-string values to boolean.
+        cols = np.array(fitstable.dtype.names)
+        boolcols = cols[np.hstack(fitstable.dtype.descr)[1::2] == '<U5']
+        for col in boolcols:
+            fitstable[col] = fitstable[col] == 'true'
+
+        # ADM only write out the columns we need for targeting.
+        nobjs = len(fitstable)
+        done = np.zeros(nobjs, dtype=ingaiadatamodel.dtype)
+        for col in done.dtype.names:
+            if col == 'REF_CAT':
+                done[col] = 'G2'
+            else:
+                done[col] = fitstable[col.lower()]
+        fitsio.write(outfile, done, extname='GAIAFITS')
+
+        # ADM return the HEALPixels that this file touches.
+        pix = set(radec2pix(nside, fitstable["ra"], fitstable["dec"]))
+        return [pix, os.path.basename(outfile)]
+
+    # ADM this is just to count processed files in _update_status.
+    nfile = np.zeros((), dtype='i8')
+    t0 = time()
+
+    def _update_status(result):
+        """wrapper function for the critical reduction operation,
+        that occurs on the main parallel process"""
+        if nfile % 100 == 0 and nfile > 0:
+            rate = nfile / (time() - t0)
+            elapsed = time() - t0
+            log.info(
+                '{}/{} files; {:.1f} files/sec; {:.1f} total mins elapsed'
+                .format(nfile, nfiles, rate, elapsed/60.)
+            )
+        nfile[...] += 1    # this is an in-place modification
+        return result
+
+    # - Parallel process input files...
+    if numproc > 1:
+        pool = sharedmem.MapReduce(np=numproc)
+        with pool:
+            pixinfile = pool.map(_write_gaia_fits, infiles, reduce=_update_status)
+    # ADM ...or run in serial.
+    else:
+        pixinfile = list()
+        for file in infiles:
+            pixinfile.append(_update_status(_write_gaia_fits(file)))
+
+    # ADM create a list for which each index is a HEALPixel and each
+    # ADM entry is a list of files that touch that HEALPixel.
+    npix = hp.nside2npix(nside)
+    pixlist = [[] for i in range(npix)]
+    for pixels, file in pixinfile:
+        for pix in pixels:
+            pixlist[pix].append(file)
+
+    # ADM write out the HEALPixel->files look-up table.
+    outfilename = os.path.join(fitsdir, "hpx-to-files.pickle")
+    outfile = open(outfilename, "wb")
+    pickle.dump(pixlist, outfile)
+    outfile.close()
+
+    log.info('Done...t={:.1f}s'.format(time()-t0))
+
+    return
+
+
+def gaia_fits_to_healpix(numproc=4):
+    """Convert files in $GAIA_DIR/fits to files in $GAIA_DIR/healpix.
+
+    Parameters
+    ----------
+    numproc : :class:`int`, optional, defaults to 4
+        The number of parallel processes to use.
+
+    Returns
+    -------
+    Nothing
+        But the archived Gaia FITS files in $GAIA_DIR/fits are
+        rearranged by HEALPixel in the directory $GAIA_DIR/healpix.
+        The HEALPixel sense is nested with nside=_get_gaia_nside(), and
+        each file in $GAIA_DIR/healpix is called healpy-xxxxx.fits,
+        where xxxxx corresponds to the HEALPixel number.
+
+    Notes
+    -----
+        - The environment variable $GAIA_DIR must be set.
+        - if numproc==1, use the serial code instead of the parallel code.
+        - Runs in about 1-3 hours with numproc=32 for 60,000 files.
+    """
+    # ADM the resolution at which the Gaia HEALPix files should be stored.
+    nside = _get_gaia_nside()
+
+    # ADM check that the GAIA_DIR is set.
+    gaiadir = _get_gaia_dir()
+
+    # ADM construct the directories for reading/writing files.
+    fitsdir = os.path.join(gaiadir, 'fits')
+    hpxdir = os.path.join(gaiadir, 'healpix')
+
+    # ADM make sure the output directory is empty.
+    if os.path.exists(hpxdir):
+        if len(os.listdir(hpxdir)) > 0:
+            msg = "{} should be empty to make Gaia HEALPix files!".format(hpxdir)
+            log.critical(msg)
+            raise ValueError(msg)
+    # ADM make the output directory, if needed.
+    else:
+        log.info('Making Gaia directory for storing HEALPix files')
+        os.makedirs(hpxdir)
+
+    # ADM read the pixel -> file look-up table.
+    infilename = os.path.join(fitsdir, "hpx-to-files.pickle")
+    infile = open(infilename, "rb")
+    pixlist = pickle.load(infile)
+    npixels = len(pixlist)
+    # ADM include the pixel number explicitly in the look-up table.
+    pixlist = list(zip(np.arange(npixels), pixlist))
+
+    # ADM the critical function to run on every file.
+    def _write_hpx_fits(pixlist):
+        """from files that touch a pixel, write out objects in each pixel"""
+        pixnum, files = pixlist
+        # ADM only proceed if some files touch a pixel.
+        if len(files) > 0:
+            # ADM track if it's our first time through the files loop.
+            first = True
+            # ADM Read in files that touch a pixel.
+            for file in files:
+                filename = os.path.join(fitsdir, file)
+                objs = fitsio.read(filename)
+                # ADM only retain objects in the correct pixel.
+                pix = radec2pix(nside, objs["RA"], objs["DEC"])
+                if first:
+                    done = objs[pix == pixnum]
+                    first = False
+                else:
+                    done = np.hstack([done, objs[pix == pixnum]])
+            # ADM construct the name of the output file.
+            outfilename = 'healpix-{:05d}.fits'.format(pixnum)
+            outfile = os.path.join(hpxdir, outfilename)
+            # ADM write out the file.
+            hdr = fitsio.FITSHDR()
+            hdr['HPXNSIDE'] = nside
+            hdr['HPXNEST'] = True
+            fitsio.write(outfile, done, extname='GAIAHPX', header=hdr)
+
+        return
+
+    # ADM this is just to count processed files in _update_status.
+    npix = np.zeros((), dtype='i8')
+    t0 = time()
+
+    def _update_status(result):
+        """wrapper function for the critical reduction operation,
+        that occurs on the main parallel process"""
+        if npix % 100 == 0 and npix > 0:
+            rate = npix / (time() - t0)
+            elapsed = time() - t0
+            log.info(
+                '{}/{} files; {:.1f} files/sec; {:.1f} total mins elapsed'
+                .format(npix, npixels, rate, elapsed/60.)
+            )
+        npix[...] += 1    # this is an in-place modification
+        return result
+
+    # - Parallel process input files...
+    if numproc > 1:
+        pool = sharedmem.MapReduce(np=numproc)
+        with pool:
+            _ = pool.map(_write_hpx_fits, pixlist, reduce=_update_status)
+    # ADM ...or run in serial.
+    else:
+        for pix in pixlist:
+            _update_status(_write_hpx_fits(pix))
+
+    log.info('Done...t={:.1f}s'.format(time()-t0))
+
+    return
+
+
+def make_gaia_files(numproc=4, download=False):
+    """Make the HEALPix-split Gaia DR2 files used by desitarget.
+
+    Parameters
+    ----------
+    numproc : :class:`int`, optional, defaults to 4
+        The number of parallel processes to use.
+    download : :class:`bool`, optional, defaults to ``False``
+        If ``True`` then wget the Gaia DR2 csv files from ESA.
+
+    Returns
+    -------
+    Nothing
+        But produces:
+        - Full Gaia DR2 CSV files in $GAIA_DIR/csv.
+        - FITS files with columns from `ingaiadatamodel` in $GAIA_DIR/fits.
+        - FITS files reorganized by HEALPixel in $GAIA_DIR/healpix.
+
+        The HEALPixel sense is nested with nside=_get_gaia_nside(), and
+        each file in $GAIA_DIR/healpix is called healpy-xxxxx.fits,
+        where xxxxx corresponds to the HEALPixel number.
+
+    Notes
+    -----
+        - The environment variable $GAIA_DIR must be set.
+        - if numproc==1, use the serial code instead of the parallel code.
+        - Runs in about 26 hours if download is ``True``.
+        - Runs in 1-3 hours with numproc=32 if download is ``False``.
+    """
+    t0 = time()
+    log.info('Begin making Gaia files...t={:.1f}s'.format(time()-t0))
+
+    # ADM check that the GAIA_DIR is set.
+    gaiadir = _get_gaia_dir()
+
+    # ADM a quick check that the fits and healpix directories are empty
+    # ADM before embarking on the slower parts of the code.
+    fitsdir = os.path.join(gaiadir, 'fits')
+    hpxdir = os.path.join(gaiadir, 'healpix')
+    for direc in [fitsdir, hpxdir]:
+        if os.path.exists(direc):
+            if len(os.listdir(direc)) > 0:
+                msg = "{} should be empty to make Gaia files!".format(direc)
+                log.critical(msg)
+                raise ValueError(msg)
+
+    if download:
+        scrape_gaia()
+        log.info('Retrieved Gaia files from ESA...t={:.1f}s'.format(time()-t0))
+
+    gaia_csv_to_fits(numproc=numproc)
+    log.info('Converted CSV files to FITS...t={:.1f}s'.format(time()-t0))
+
+    gaia_fits_to_healpix(numproc=numproc)
+    log.info('Rearranged FITS files by HEALPixel...t={:.1f}s'.format(time()-t0))
+
+    return
 
 
 def pop_gaia_coords(inarr):
@@ -66,7 +474,7 @@ def pop_gaia_coords(inarr):
     names = list(inarr.dtype.names)
 
     # ADM pop off any instances of GAIA_RA, GAIA_DEC be forgiving if they
-    # ADM aren't in the array
+    # ADM aren't in the array.
     try:
         names.remove("GAIA_RA")
     except ValueError:
@@ -81,8 +489,38 @@ def pop_gaia_coords(inarr):
     return inarr[names]
 
 
+def pop_gaia_columns(inarr, cols):
+    """Convenience function to pop columns of an input array.
+
+    Parameters
+    ----------
+    inarr : :class:`numpy.ndarray`
+        Structured array with various column names.
+    cols : :class:`list`
+        List of columns to remove from the input array.
+
+    Returns
+    -------
+    :class:`numpy.ndarray`
+        Input array with columns in cols removed.
+    """
+    # ADM list of the column names of the passed array
+    names = list(inarr.dtype.names)
+
+    # ADM pop off any instances of GAIA_RA, GAIA_DEC be forgiving if they
+    # ADM aren't in the array.
+    for col in cols:
+        try:
+            names.remove(col)
+        except ValueError:
+            pass
+
+    # ADM return the array without GAIA_RA, GAIA_DEC
+    return inarr[names]
+
+
 def read_gaia_file(filename, header=False):
-    """Read in a Gaia "chunks" file in the appropriate format for desitarget
+    """Read in a Gaia healpix file in the appropriate format for desitarget.
 
     Parameters
     ----------
@@ -96,7 +534,7 @@ def read_gaia_file(filename, header=False):
     -------
     :class:`numpy.ndarray`
         Gaia data translated to targeting format (upper-case etc.) with the
-        columns corresponding to `desitarget.secondary.gaiadatamodel`
+        columns corresponding to `desitarget.gaiamatch.gaiadatamodel`
 
     Notes
     -----
@@ -132,9 +570,8 @@ def read_gaia_file(filename, header=False):
         return outdata
 
 
-def find_gaia_files(objs, neighbors=True,
-                    gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
-    """Find full paths to all relevant Gaia "chunks" files for an object array
+def find_gaia_files(objs, neighbors=True):
+    """Find full paths to all relevant Gaia healpix files for an object array.
 
     Parameters
     ----------
@@ -144,32 +581,38 @@ def find_gaia_files(objs, neighbors=True,
         Return all of the pixels that touch the Gaia files of interest
         in order to prevent edge effects (e.g. if a Gaia source is 1 arcsec
         away from a primary source and so in an adjacent pixel)
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
 
     Returns
     -------
     :class:`list`
         A list of all Gaia files that need to be read in to account for objects
         at the passed locations.
-    """
-    # ADM the resolution at which the chunks files are stored
-    nside = 32
 
-    # ADM convert RA/Dec to co-latitude and longitude in radians; note that the
-    # ADM Legacy Surveys do NOT use the NESTED scheme for storing Gaia files
+    Notes
+    -----
+        - The environment variable $GAIA_DIR must be set.
+    """
+    # ADM the resolution at which the Gaia HEALPix files are stored.
+    nside = _get_gaia_nside()
+
+    # ADM check that the GAIA_DIR is set and retrieve it.
+    gaiadir = _get_gaia_dir()
+    hpxdir = os.path.join(gaiadir, 'healpix')
+
+    # ADM convert RA/Dec to co-latitude and longitude in radians.
     theta, phi = np.radians(90-objs["DEC"]), np.radians(objs["RA"])
 
-    # ADM retrieve the pixels in which the locations lie
-    pixnum = hp.ang2pix(nside, theta, phi)
+    # ADM retrieve the pixels in which the locations lie.
+    pixnum = hp.ang2pix(nside, theta, phi, nest=True)
     # ADM if neighbors was sent, then retrieve all pixels that touch each
     # ADM pixel covered by the provided locations, to prevent edge effects...
     if neighbors:
-        pixnum = np.hstack([pixnum,
-                            np.hstack(hp.pixelfunc.get_all_neighbours(nside, theta, phi))])
+        pixnum = np.hstack(
+            [pixnum, np.hstack(hp.pixelfunc.get_all_neighbours(nside, theta, phi, nest=True))]
+        )
 
     # ADM retrieve only the UNIQUE pixel numbers. It's possible that only
-    # ADM one pixel was produced, so guard against pixnum being non-iterable
+    # ADM one pixel was produced, so guard against pixnum being non-iterable.
     if not isinstance(pixnum, np.integer):
         pixnum = list(set(pixnum))
     else:
@@ -179,15 +622,14 @@ def find_gaia_files(objs, neighbors=True,
     if -1 in pixnum:
         pixnum.remove(-1)
 
-    # ADM reformat in the Gaia chunked format used by the Legacy Surveys
-    gaiafiles = ['{}/chunk-{:05d}.fits'.format(gaiadir, pn) for pn in pixnum]
+    # ADM reformat in the Gaia healpix format used by desitarget.
+    gaiafiles = [os.path.join(hpxdir, 'healpix-{:05d}.fits'.format(pn)) for pn in pixnum]
 
     return gaiafiles
 
 
-def find_gaia_files_box(gaiabounds, neighbors=True,
-                        gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
-    """Find full paths to all relevant gaia "chunks" files for an object array
+def find_gaia_files_box(gaiabounds, neighbors=True):
+    """Find full paths to all relevant Gaia healpix files in an RA/Dec box.
 
     Parameters
     ----------
@@ -197,8 +639,6 @@ def find_gaia_files_box(gaiabounds, neighbors=True,
     neighbors : :class:`bool`, optional, defaults to ``True``
         Return all of the pixels that touch the pixels in the box in
         order to guard against edge effects
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
 
     Returns
     -------
@@ -212,9 +652,14 @@ def find_gaia_files_box(gaiabounds, neighbors=True,
           warnings about returning different pixel sets at different values
           of `fact` apply. See:
           https://healpy.readthedocs.io/en/latest/generated/healpy.query_polygon.html
+        - The environment variable $GAIA_DIR must be set.
     """
-    # ADM the resolution at which the chunks files are stored
-    nside = 32
+    # ADM the resolution at which the healpix files are stored.
+    nside = _get_gaia_nside()
+
+    # ADM check that the GAIA_DIR is set and retrieve it.
+    gaiadir = _get_gaia_dir()
+    hpxdir = os.path.join(gaiadir, 'healpix')
 
     # ADM retrive the RA/Dec bounds from the passed list
     ramin, ramax, decmin, decmax = gaiabounds
@@ -228,16 +673,17 @@ def find_gaia_files_box(gaiabounds, neighbors=True,
     # ADM transpose to pass the array to query_polygon in the correct order
     vecs = hp.dir2vec(thetapairs, phipairs).T
 
-    # ADM determine the pixels that touch the box; note that the Legacy
-    # ADM Surveys do NOT use the NESTED scheme for storing Gaia files
-    pixnum = hp.query_polygon(nside, vecs, inclusive=True, fact=4)
+    # ADM determine the pixels that touch the box.
+    pixnum = hp.query_polygon(nside, vecs, inclusive=True, fact=4, nest=True)
 
     # ADM if neighbors was sent, then retrieve all pixels that touch each
     # ADM pixel covered by the provided locations, to prevent edge effects...
     if neighbors:
-        # ADM first convert back to theta/phi to retrieve neighbors
-        theta, phi = hp.pix2ang(nside, pixnum)
-        pixnum = np.hstack(hp.pixelfunc.get_all_neighbours(nside, theta, phi))
+        # ADM first convert back to theta/phi to retrieve neighbors.
+        theta, phi = hp.pix2ang(nside, pixnum, nest=True)
+        pixnum = np.hstack(
+            hp.pixelfunc.get_all_neighbours(nside, theta, phi, nest=True)
+        )
 
     # ADM retrieve only the UNIQUE pixel numbers. It's possible that only
     # ADM one pixel was produced, so guard against pixnum being non-iterable
@@ -250,16 +696,15 @@ def find_gaia_files_box(gaiabounds, neighbors=True,
     if -1 in pixnum:
         pixnum.remove(-1)
 
-    # ADM format in the gaia chunked format used by the Legacy Surveys
-    gaiafiles = ['{}/chunk-{:05d}.fits'.format(gaiadir, pn) for pn in pixnum]
+    # ADM reformat in the Gaia healpix format used by desitarget.
+    gaiafiles = [os.path.join(hpxdir, 'healpix-{:05d}.fits'.format(pn)) for pn in pixnum]
 
     return gaiafiles
 
 
-def match_gaia_to_primary(objs, matchrad=1.,
-                          retaingaia=False, gaiabounds=[0., 360., -90., 90.],
-                          gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
-    """Match a set of objects to Gaia "chunks" files and return the Gaia information
+def match_gaia_to_primary(objs, matchrad=1., retaingaia=False,
+                          gaiabounds=[0., 360., -90., 90.]):
+    """Match a set of objects to Gaia healpix files and return the Gaia information.
 
     Parameters
     ----------
@@ -277,8 +722,6 @@ def match_gaia_to_primary(objs, matchrad=1.,
         Used in conjunction with `retaingaia` to determine over what area to
         retrieve Gaia objects that don't match a sweeps object. Pass a 4-entry
         list to represent an area bounded by [RAmin, RAmax, DECmin, DECmax]
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
 
     Returns
     -------
@@ -302,35 +745,35 @@ def match_gaia_to_primary(objs, matchrad=1.,
     # ADM e.g. https://stackoverflow.com/questions/40845304/runtimewarning-numpy-dtype-size-changed-may-indicate-binary-incompatibility
     import warnings
 
-    # ADM if retaingaia is True, retain all Gaia objects in a sweeps-like box
+    # ADM if retaingaia is True, retain all Gaia objects in a sweeps-like box.
     if retaingaia:
         ramin, ramax, decmin, decmax = gaiabounds
 
-    # ADM convert the coordinates of the input objects to a SkyCoord object
+    # ADM convert the coordinates of the input objects to a SkyCoord object.
     cobjs = SkyCoord(objs["RA"]*u.degree, objs["DEC"]*u.degree)
     nobjs = cobjs.size
 
-    # ADM deal with the special case that only a single object was passed
+    # ADM deal with the special case that only a single object was passed.
     if nobjs == 1:
-        return match_gaia_to_primary_single(objs, matchrad=matchrad, gaiadir=gaiadir)
+        return match_gaia_to_primary_single(objs, matchrad=matchrad)
 
-    # ADM set up a zerod array of Gaia information for the passed objects
+    # ADM set up a zerod array of Gaia information for the passed objects.
     gaiainfo = np.zeros(nobjs, dtype=gaiadatamodel.dtype)
 
     # ADM a supplemental (zero-length) array to hold Gaia objects that don't
-    # ADM match a sweeps object, in case retaingaia was set
+    # ADM match a sweeps object, in case retaingaia was set.
     suppgaiainfo = np.zeros(0, dtype=gaiadatamodel.dtype)
 
-    # ADM objects without matches should have REF_ID of -1
+    # ADM objects without matches should have REF_ID of -1.
     gaiainfo['REF_ID'] = -1
 
-    # ADM determine which Gaia files need to be considered
+    # ADM determine which Gaia files need to be considered.
     if retaingaia:
-        gaiafiles = find_gaia_files_box(gaiabounds, gaiadir=gaiadir)
+        gaiafiles = find_gaia_files_box(gaiabounds)
     else:
-        gaiafiles = find_gaia_files(objs, gaiadir=gaiadir)
+        gaiafiles = find_gaia_files(objs)
 
-    # ADM loop through the Gaia files and match to the passed objects
+    # ADM loop through the Gaia files and match to the passed objects.
     for file in gaiafiles:
         gaia = read_gaia_file(file)
         cgaia = SkyCoord(gaia["GAIA_RA"]*u.degree, gaia["GAIA_DEC"]*u.degree)
@@ -338,23 +781,23 @@ def match_gaia_to_primary(objs, matchrad=1.,
             warnings.simplefilter("ignore")
             # ADM ****here's where the warning occurs...
             idobjs, idgaia, _, _ = cgaia.search_around_sky(cobjs, matchrad*u.arcsec)
-        # ADM assign the Gaia info to the array that corresponds to the passed objects
+        # ADM assign the Gaia info to the array that corresponds to the passed objects.
         gaiainfo[idobjs] = gaia[idgaia]
 
         # ADM if retaingaia was set, also build an array of Gaia objects that
-        # ADM don't have sweeps matches, but are within the RA/Dec bounds
+        # ADM don't have sweeps matches, but are within the RA/Dec bounds.
         if retaingaia:
-            # ADM find the Gaia IDs that didn't match the passed objects
+            # ADM find the Gaia IDs that didn't match the passed objects.
             nomatch = set(np.arange(len(gaia)))-set(idgaia)
             noidgaia = np.array(list(nomatch))
-            # ADM which Gaia objects with these IDs are within the bounds
+            # ADM which Gaia objects with these IDs are within the bounds.
             if len(noidgaia) > 0:
                 suppg = gaia[noidgaia]
                 winbounds = np.where(
                     (suppg["GAIA_RA"] >= ramin) & (suppg["GAIA_RA"] < ramax)
                     & (suppg["GAIA_DEC"] >= decmin) & (suppg["GAIA_DEC"] < decmax)
                 )[0]
-                # ADM Append those Gaia objects to the suppgaiainfo array
+                # ADM Append those Gaia objects to the suppgaiainfo array.
                 if len(winbounds) > 0:
                     suppgaiainfo = np.hstack([suppgaiainfo, suppg[winbounds]])
 
@@ -364,9 +807,8 @@ def match_gaia_to_primary(objs, matchrad=1.,
     return gaiainfo
 
 
-def match_gaia_to_primary_single(objs, matchrad=1.,
-                                 gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
-    """Match ONE object to Gaia "chunks" files and return the Gaia information
+def match_gaia_to_primary_single(objs, matchrad=1.):
+    """Match ONE object to Gaia "chunks" files and return the Gaia information.
 
     Parameters
     ----------
@@ -374,8 +816,6 @@ def match_gaia_to_primary_single(objs, matchrad=1.,
         Must contain at least "RA" and "DEC". MUST BE A SINGLE ROW.
     matchrad : :class:`float`, optional, defaults to 1 arcsec
         The matching radius in arcseconds.
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
 
     Returns
     -------
@@ -395,22 +835,22 @@ def match_gaia_to_primary_single(objs, matchrad=1.,
     # ADM e.g. https://stackoverflow.com/questions/40845304/runtimewarning-numpy-dtype-size-changed-may-indicate-binary-incompatibility
     import warnings
 
-    # ADM convert the coordinates of the input objects to a SkyCoord object
+    # ADM convert the coordinates of the input objects to a SkyCoord object.
     cobjs = SkyCoord(objs["RA"]*u.degree, objs["DEC"]*u.degree)
     nobjs = cobjs.size
     if nobjs > 1:
         log.error("Only matches one row but {} rows were sent".format(nobjs))
 
-    # ADM set up a zerod array of Gaia information for the passed object
+    # ADM set up a zerod array of Gaia information for the passed object.
     gaiainfo = np.zeros(nobjs, dtype=gaiadatamodel.dtype)
 
-    # ADM an object without matches should have REF_ID of -1
+    # ADM an object without matches should have REF_ID of -1.
     gaiainfo['REF_ID'] = -1
 
-    # ADM determine which Gaia files need to be considered
-    gaiafiles = find_gaia_files(objs, gaiadir=gaiadir)
+    # ADM determine which Gaia files need to be considered.
+    gaiafiles = find_gaia_files(objs)
 
-    # ADM loop through the Gaia files and match to the passed object
+    # ADM loop through the Gaia files and match to the passed object.
     for file in gaiafiles:
         gaia = read_gaia_file(file)
         cgaia = SkyCoord(gaia["GAIA_RA"]*u.degree, gaia["GAIA_DEC"]*u.degree)
@@ -419,15 +859,14 @@ def match_gaia_to_primary_single(objs, matchrad=1.,
             # ADM ****here's where the warning occurs...
             sep = cobjs.separation(cgaia)
             idgaia = np.where(sep < matchrad*u.arcsec)[0]
-        # ADM assign the Gaia info to the array that corresponds to the passed object
+        # ADM assign the Gaia info to the array that corresponds to the passed object.
         if len(idgaia) > 0:
             gaiainfo = gaia[idgaia]
 
     return gaiainfo
 
 
-def write_gaia_matches(infiles, numproc=4, outdir=".",
-                       gaiadir='/project/projectdirs/cosmo/work/gaia/chunks-gaia-dr2-astrom'):
+def write_gaia_matches(infiles, numproc=4, outdir="."):
     """Match sweeps files to Gaia and rewrite with the Gaia columns added
 
     Parameters
@@ -439,8 +878,6 @@ def write_gaia_matches(infiles, numproc=4, outdir=".",
         The number of parallel processes to use.
     outdir : :class:`str`, optional, default to the current directory
         The directory to write the files.
-    gaiadir : :class:`str`, optional, defaults to Gaia DR2 path at NERSC
-        Root directory of a Gaia Data Release as used by the Legacy Surveys.
 
     Returns
     -------
@@ -449,62 +886,62 @@ def write_gaia_matches(infiles, numproc=4, outdir=".",
         added (except for the columns `GAIA_RA` and `GAIA_DEC`) are
         written to file. The filename is the same as the input
         filename with the ".fits" replaced by "-gaia$DRmatch.fits"
-        where $DR is extracted from the passed gaiadir.
+        where $DR is extracted from the $GAIA_DIR environment variable.
 
     Notes
     -----
         - if numproc==1, use the serial code instead of the parallel code.
+        - The environment variable $GAIA_DIR must be set.
     """
-    from os.path import basename
-    from desitarget import io
-    from desitarget.internal import sharedmem
+    # ADM check that the GAIA_DIR is set and retrieve it.
+    gaiadir = _get_gaia_dir()
 
-    # ADM convert a single file, if passed to a list of files
+    # ADM convert a single file, if passed to a list of files.
     if isinstance(infiles, str):
         infiles = [infiles, ]
 
-    # ADM check that files exist before proceeding
+    # ADM check that files exist before proceeding.
     for filename in infiles:
         if not os.path.exists(filename):
             raise ValueError("{} doesn't exist".format(filename))
 
     nfiles = len(infiles)
 
-    # ADM extract a reasonable name for output files from the Gaia directory
+    # ADM extract a reasonable name for output files from the Gaia directory.
     drloc = gaiadir.find("dr")
-    # ADM if we didn't find the substring "dr" go generic
+    # ADM if we didn't find the substring "dr" go generic.
     if drloc == -1:
         ender = '-gaiamatch.fits'
     else:
         ender = '-gaia{}match.fits'.format(gaiadir[drloc:drloc+3])
 
-    # ADM the critical function to run on every file
+    # ADM the critical function to run on every file.
     def _get_gaia_matches(fnwdir):
         '''wrapper on match_gaia_to_primary() given a file name'''
-        # ADM extract the output file name
-        fn = basename(fnwdir)
+        # ADM extract the output file name.
+        fn = os.path.basename(fnwdir)
         outfile = '{}/{}'.format(outdir, fn.replace(".fits", ender))
 
-        # ADM read in the objects
+        # ADM read in the objects.
         objs, hdr = io.read_tractor(fnwdir, header=True)
 
-        # ADM match to Gaia sources
-        gaiainfo = match_gaia_to_primary(objs, gaiadir=gaiadir)
+        # ADM match to Gaia sources.
+        gaiainfo = match_gaia_to_primary(objs)
         log.info('Done with Gaia match for {} primary objects...t = {:.1f}s'
                  .format(len(objs), time()-start))
 
         # ADM remove the GAIA_RA, GAIA_DEC columns as they aren't
-        # ADM in the imaging surveys data model
+        # ADM in the imaging surveys data model.
         gaiainfo = pop_gaia_coords(gaiainfo)
 
-        # ADM add the Gaia column information to the sweeps array
+        # ADM add the Gaia column information to the sweeps array.
         for col in gaiainfo.dtype.names:
             objs[col] = gaiainfo[col]
 
         fitsio.write(outfile, objs, extname='SWEEP', header=hdr, clobber=True)
         return True
 
-    # ADM this is just to count sweeps files in _update_status
+    # ADM this is just to count sweeps files in _update_status.
     nfile = np.zeros((), dtype='i8')
 
     t0 = time()
@@ -515,10 +952,10 @@ def write_gaia_matches(infiles, numproc=4, outdir=".",
         if nfile % 50 == 0 and nfile > 0:
             rate = nfile / (time() - t0)
             log.info('{}/{} files; {:.1f} files/sec'.format(nfile, nfiles, rate))
-        nfile[...] += 1    # this is an in-place modification
+        nfile[...] += 1    # this is an in-place modification.
         return result
 
-    # - Parallel process input files
+    # - Parallel process input files.
     if numproc > 1:
         pool = sharedmem.MapReduce(np=numproc)
         with pool: